# api/openapi3.yaml
openapi: 3.0.0
info:
  contact:
    email: support@wishlist.example.com
    name: API Support
  description: >-
    A RESTful API for managing wish lists, gift items, and reservations.

    Features include user authentication, wish list management, gift item tracking, and reservation system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Wish List API
  version: "1.1"
paths:
...
  /items:
    get:
      description: Get all gift items owned by the authenticated user with pagination
        and filters
      parameters:
        - description: Page number (default 1)
          in: query
          name: page
          schema:
            type: integer
        - description: Items per page (default 10, max 100)
          in: query
          name: limit
          schema:
            type: integer
        - description: Sort field (created_at, updated_at, title, price)
          in: query
          name: sort
          schema:
            type: string
        - description: Sort order (asc, desc)
          in: query
          name: order
          schema:
            type: string
        - description: Filter items not attached to any wishlist
          in: query
          name: unattached
          schema:
            type: boolean
        - description: Include archived items (default false)
          in: query
          name: include_archived
          schema:
            type: boolean
        - description: Search in title and description
          in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of items retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.Pa\
                  ginatedItemsResponse"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get my gift items
      tags:
        - Items
    post:
      description: Create a new gift item without attaching it to a wishlist
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.Cr\
                eateItemRequest"
        description: Item data
        required: true
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.It\
                  emResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Create gift item
      tags:
        - Items
  "/items/{id}":
    delete:
      description: Archive a gift item by setting archived_at timestamp. Item is
        removed from all queries but data is preserved.
      parameters:
        - description: Item ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Item archived successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Delete gift item (soft delete)
      tags:
        - Items
    get:
      description: Get a specific gift item by ID
      parameters:
        - description: Item ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.It\
                  emResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get gift item
      tags:
        - Items
    put:
      description: Update a gift item by ID
      parameters:
        - description: Item ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.Up\
                dateItemRequest"
        description: Updated item data
        required: true
      responses:
        "200":
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.It\
                  emResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Update gift item
      tags:
        - Items
  "/items/{id}/mark-purchased":
    post:
      description: Mark a gift item as purchased with the actual purchased price. This
        is a global status.
      parameters:
        - description: Item ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.Ma\
                rkPurchasedRequest"
        description: Purchase details
        required: true
      responses:
        "200":
          description: Item marked as purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_item_delivery_http_dto.It\
                  emResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Mark gift item as purchased
      tags:
        - Items
...
servers:
  - url: //localhost:8080/api
components:
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...